fix last commit

+++ proxy/plugin/solana_rest_api_tools.py
@@ -107,7 +107,7 @@ def estimate_gas(tx_sender: NeonTxSender, sender,  *, logger):
         tx_sender.steps_emulated += EVM_STEPS
 
     if tx_sender.steps_emulated > 0:
-        full_step_iterations = tx_sender.steps_emulated / evm_steps_by_trx
+        full_step_iterations = int(tx_sender.steps_emulated / evm_steps_by_trx)
         final_steps =  tx_sender.steps_emulated % evm_steps_by_trx
         if final_steps > 0 and final_steps < EVM_STEPS:
             final_steps = EVM_STEPS
